PAI_Lab/Lab-10/requirements.txt
weaviate-client
streamlit
openai
pandas
dotenv
